Add support for Rails 4.2
We are adding support for Rails 4.2 as there are still services using
this version. But we can also phase out this support if we find
ourselves limited in the future.

Since Rails 4.2 has a different way of dealing with primary keys, we
keep the responsibility of defining the primary key outside the type
definition. As this is not what the gem is expected to do.