Remove primary_key option from type declaration
When working previously on Rails 4.2 support, we realised that the
`primary_key` option in type declation model can be confusing.
One can easily get tricked into thinking that this will override what's
in the schema and declare the column as a primary key.

This option is only used to decide if we should automatically fill the column or
not.

To remove the confusion, we will keep only the `auto_generate` option
and discard the `primary_key` one.