CI: Fix events triggering lint/test jobs
With the old configuration, we would run CI twice when pushing a commit and
opening a pull request. Or when pushing a commit that belongs to a pull request.

With the new configuration, we run CI once when opening a PR, and once when
pushing a commit that belongs to a pull request. And once when merging to
master (the `push` event). Importantly, this configuration still works for
contributions from forked repositories, since the `pull_request` events are
triggered on _our_ fork.