Add many more tests
I don't really have confidence in the singular `test_eager_loading` test, so
adding tests that checks use of `includes`, `joins`, `left_outer_joins` and
`preload` too.

We also added tests to ensure we can read the format we have been writing to
the database so far, and that we write to the database in the same format as
usual.